Extracted the 'ReadConversation' and 'WriteConversation' functions into distinct classes to adhere to the Single Responsibility principle
