+++ MyChat/ConversationExporter.cs
@@ -1,13 +1,8 @@
 ﻿namespace MyChat
 {
     using System;
-    using System.Collections.Generic;
-    using System.IO;
-    using System.Security;
-    using System.Text;
     using Microsoft.Extensions.Configuration;
     using MindLink.Recruitment.MyChat;
-    using Newtonsoft.Json;
 
     /// <summary>
     /// Represents a conversation exporter that can read a conversation and write it out in JSON.
@@ -47,105 +42,14 @@
         /// </exception>
         public void ExportConversation(string inputFilePath, string outputFilePath)
         {
-            Conversation conversation = this.ReadConversation(inputFilePath);
+            ConversationReader conversationReader = new ConversationReader();
+            ConversationWriter conversationWriter = new ConversationWriter();
 
-            this.WriteConversation(conversation, outputFilePath);
+            Conversation conversation = conversationReader.ReadConversation(inputFilePath);
+
+            conversationWriter.WriteConversation(conversation, outputFilePath);
 
             Console.WriteLine("Conversation exported from '{0}' to '{1}'", inputFilePath, outputFilePath);
-        }
-
-        /// <summary>
-        /// Helper method to read the conversation from <paramref name="inputFilePath"/>.
-        /// </summary>
-        /// <param name="inputFilePath">
-        /// The input file path.
-        /// </param>
-        /// <returns>
-        /// A <see cref="Conversation"/> model representing the conversation.
-        /// </returns>
-        /// <exception cref="ArgumentException">
-        /// Thrown when the input file could not be found.
-        /// </exception>
-        /// <exception cref="Exception">
-        /// Thrown when something else went wrong.
-        /// </exception>
-        public Conversation ReadConversation(string inputFilePath)
-        {
-            try
-            {
-                var reader = new StreamReader(new FileStream(inputFilePath, FileMode.Open, FileAccess.Read),
-                    Encoding.ASCII);
-
-                string conversationName = reader.ReadLine();
-                var messages = new List<Message>();
-
-                string line;
-                
-                const char space = ' ';
-
-                while ((line = reader.ReadLine()) != null)
-                {
-                    var split = line.Split(space);
-
-                    // Fixes bug by using array range operator
-                    var content = string.Join(space, split[2..split.Length]);
-
-                    messages.Add(new Message(DateTimeOffset.FromUnixTimeSeconds(Convert.ToInt64(split[0])), split[1], content));
-                }
-
-                return new Conversation(conversationName, messages);
-            }
-            catch (FileNotFoundException)
-            {
-                throw new ArgumentException("The file was not found.");
-            }
-            catch (IOException)
-            {
-                throw new Exception("Something went wrong in the IO.");
-            }
-        }
-
-        /// <summary>
-        /// Helper method to write the <paramref name="conversation"/> as JSON to <paramref name="outputFilePath"/>.
-        /// </summary>
-        /// <param name="conversation">
-        /// The conversation.
-        /// </param>
-        /// <param name="outputFilePath">
-        /// The output file path.
-        /// </param>
-        /// <exception cref="ArgumentException">
-        /// Thrown when there is a problem with the <paramref name="outputFilePath"/>.
-        /// </exception>
-        /// <exception cref="Exception">
-        /// Thrown when something else bad happens.
-        /// </exception>
-        public void WriteConversation(Conversation conversation, string outputFilePath)
-        {
-            try
-            {
-                var writer = new StreamWriter(new FileStream(outputFilePath, FileMode.Create, FileAccess.ReadWrite));
-
-                var serialized = JsonConvert.SerializeObject(conversation, Formatting.Indented);
-
-                writer.Write(serialized);
-
-                writer.Flush();
-
-                writer.Close();
-            }
-            catch (SecurityException)
-            {
-                throw new ArgumentException("No permission to file.");
-            }
-            catch (DirectoryNotFoundException)
-            {
-                throw new ArgumentException("Path invalid.");
-            }
-            catch (IOException)
-            {
-                throw new Exception("Something went wrong in the IO.");
-            }
         }
     }
 }
